# express_tsoa_template/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					}
				},
				"required": [
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"in": "header",
				"name": "Authorization",
				"description": "Berer token to access these api",
				"scheme": "bearer",
				"bearerFormat": "JWT "
			}
		}
	},
	"info": {
		"title": "project_central_backend_api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Amit Kumar ",
			"email": "hey.amit@outlook.com"
		}
	},
	"paths": {
...
		"/api/v1/events/accept-user-signup-and-update": {
			"post": {
				"operationId": "UserSignupAndUserUpdateEventPublisher",
				"responses": {
					"200": {
						"description": "A boolean indicating success",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					},
					"401": {
						"description": "- The response object for error handling",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					},
					"403": {
						"description": "- The response object for error handling",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Handles user signup and update events from Casdoor.",
				"summary": "Save user signed up with Casdoor to the comment DB.",
				"tags": [
					"events"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The internal service secret header",
						"in": "header",
						"name": "x-internal-service",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "- The user signup or update payload",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "- The user signup or update payload"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}